+++ packages/Webkul/Installer/src/Database/Seeders/Attribute/AttributeGroupTableSeeder.php
@@ -4,6 +4,7 @@ namespace Webkul\Installer\Database\Seeders\Attribute;
 
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Schema;
 
 class AttributeGroupTableSeeder extends Seeder
 {
@@ -15,6 +16,11 @@ class AttributeGroupTableSeeder extends Seeder
      */
     public function run($parameters = [])
     {
+        // Attribute module is disabled, skip this seeder
+        if (! Schema::hasTable('attribute_groups')) {
+            return;
+        }
+
         DB::statement('SET FOREIGN_KEY_CHECKS=0;');
 
         DB::table('attribute_groups')->delete();
